+++ apps/operator/src/components/layouts/auth/auth.styles.tsx
@@ -8,7 +8,7 @@ const authStyles = tv({
     bg: 'absolute h-full w-2/3 top-0 left-0',
     bgImage: 'object-cover h-full z-10',
     content:
-      'relative z-20 bg-white shadow-auth rounded-lg flex flex-col justify-center mx-auto my-auto p-6 sm:w-1/3',
+      'relative z-20 bg-white shadow-auth rounded-lg flex flex-col justify-center mx-auto my-auto py-20 px-12 w-full max-w-lg',
   },
 })
